refactor(slack): remove unused Zod schemas and update type exports in types.ts
