+++ agent-packages/packages/slack/package.json
@@ -22,8 +22,6 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@clearfeed-ai/quix-common-agent": "^0.1.0",
-    "@langchain/core": "^0.1.30",
     "@slack/web-api": "^6.9.0",
     "zod": "^3.22.2"
   }
