+++ agent-packages/packages/slack/src/schema.ts
@@ -61,52 +61,3 @@ export const getUsersParamsSchema = z.object({
 export const getUserProfileParamsSchema = z.object({
   user_id: z.string().describe('The ID of the user')
 });
-
-// Data schemas for response entities
-export const slackUserSchema = z.object({
-  id: z.string(),
-  name: z.string(),
-  profile: z.record(z.unknown())
-});
-
-export const slackChannelSchema = z.object({
-  id: z.string(),
-  name: z.string(),
-  is_channel: z.boolean().optional(),
-  is_group: z.boolean().optional(),
-  is_private: z.boolean().optional(),
-  is_archived: z.boolean().optional(),
-  is_general: z.boolean().optional(),
-  topic: z.object({ value: z.string() }).optional(),
-  purpose: z.object({ value: z.string() }).optional()
-});
-
-export const slackMessageSchema = z.object({
-  text: z.string(),
-  user: z.string(),
-  ts: z.string(),
-  thread_ts: z.string().optional(),
-  reply_count: z.number().optional(),
-  reactions: z
-    .array(
-      z.object({
-        name: z.string(),
-        count: z.number(),
-        users: z.array(z.string())
-      })
-    )
-    .optional()
-});
-
-// Type inferences from schemas
-export type ListChannelsParams = z.infer<typeof listChannelsParamsSchema>;
-export type PostMessageParams = z.infer<typeof postMessageParamsSchema>;
-export type ReplyToThreadParams = z.infer<typeof replyToThreadParamsSchema>;
-export type AddReactionParams = z.infer<typeof addReactionParamsSchema>;
-export type GetChannelHistoryParams = z.infer<typeof getChannelHistoryParamsSchema>;
-export type GetThreadRepliesParams = z.infer<typeof getThreadRepliesParamsSchema>;
-export type GetUsersParams = z.infer<typeof getUsersParamsSchema>;
-export type GetUserProfileParams = z.infer<typeof getUserProfileParamsSchema>;
-export type SlackUser = z.infer<typeof slackUserSchema>;
-export type SlackChannel = z.infer<typeof slackChannelSchema>;
-export type SlackMessage = z.infer<typeof slackMessageSchema>;
